pyproject.toml
[project]
name = "moveinsync-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.1",
    "pydantic>=2.12.4",
    "uvicorn>=0.38.0",
    "motor>=3.6.0",
    "python-dotenv>=1.0.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "apscheduler>=3.10.4",
]
